+++ changelog.md
@@ -1,5 +1,10 @@
 <h3>Changelog</h3>
 
+## 128 - December 25, 2021
+
+* Disable script when an unavailable video is detected
+* Remove some logging
+
 ## 127 - April 5, 2021
 
 * Resize the video faster by styling the `#player-wrap` element.